move to pay the bill file
