+++ screens/payment/PayTheBill.tsx
@@ -1,13 +1,64 @@
-import { Text, View } from "react-native";
+import { CompositeScreenProps } from "@react-navigation/native";
+import { NativeStackScreenProps } from "@react-navigation/native-stack";
+import { useCallback } from "react";
+import { FlatList, TouchableOpacity, View } from "react-native";
+import MaterialIcons from "react-native-vector-icons/MaterialIcons";
 
-interface PayTheBillProps {}
+import BodyText from "@/components/text/BodyText";
+import BodyContainer from "@/components/ui/BodyContainer";
+import Colors from "@/constants/color";
+import {
+  MOCKED_PAYMENTLICENSEPLATE,
+  mockedPaymentLicensePlateProps,
+} from "@/mock/mockData";
+import { AuthenticatedStackParamList, PaymentStackParamList } from "@/types";
 
-const PayTheBill: React.FC<PayTheBillProps> = () => {
+export type PayTheBillProps = CompositeScreenProps<
+  NativeStackScreenProps<PaymentStackParamList, "PayTheBill">,
+  NativeStackScreenProps<AuthenticatedStackParamList>
+>;
+
+const PayTheBill: React.FC<PayTheBillProps> = ({ navigation }) => {
+  const pressLicensePlateHandler = () => {
+    navigation.navigate("PaymentSummary");
+  };
+  const renderLicensePlateList = useCallback(
+    (item: mockedPaymentLicensePlateProps) => {
+      return (
+        <TouchableOpacity onPress={pressLicensePlateHandler}>
+          <View
+            style={{
+              flexDirection: "row",
+              justifyContent: "space-between",
+              paddingHorizontal: 10,
+              paddingVertical: 15,
+              backgroundColor: Colors.white,
+              borderWidth: 1,
+              margin: 10,
+              borderRadius: 10,
+              alignItems: "center",
+              borderColor: Colors.gray[800],
+            }}
+          >
+            <BodyText text={item.licensePlate} />
+            <MaterialIcons name="chevron-right" size={20}></MaterialIcons>
+          </View>
+        </TouchableOpacity>
+      );
+    },
+    []
+  );
   return (
-    <View>
-      <Text>Pay the bill</Text>
-    </View>
+    <BodyContainer>
+      <FlatList
+        data={MOCKED_PAYMENTLICENSEPLATE}
+        renderItem={({ item }) => {
+          return renderLicensePlateList(item);
+        }}
+        keyExtractor={(item) => item.id}
+        overScrollMode="never"
+      />
+    </BodyContainer>
   );
 };
-
 export default PayTheBill;
